Lots of changes (read below)
Removes `load_data()` from caes.py: it will be put in another file.

Removes the explicit information about the networks from `caes.py`,
and moves it to a `models` folder. The next step is to make `caes.py`
receive from the command line which `model` it should load.

Create a function called `train()` to train the caes. The reason is
that the dataset we intend to use is actually very big and it won't
fit in the memory. Different from `Keras`, `nolearn` doesn't have a
`fit_generator()`, so I am using a loop calling `partial_fit()`.

 Changes to be committed:
	modified:   caes.py
	new file:   models/__init__.py
	renamed:    caes.py -> models/model_1.py